% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_interactions.R
\name{check_is_data_package}
\alias{check_is_data_package}
\title{Check whether reference is a data package}
\usage{
check_is_data_package(reference_id, secure = TRUE, dev = FALSE)
}
\arguments{
\item{reference_id}{Integer. A 7-digit number indicating the datastore reference to be queried.}

\item{secure}{Logical. Defaults to TRUE, indicating that the user is logged on to the NPS VPN (or is accessing the resource from an NPS office). Can be set to FALSE for non-NPS users. If \code{secure} is set to FALSE, \code{dev} is automatically set to FALSE.}

\item{dev}{Logical. Defaults to FALSE, indicating that the user wishes to work with the production version of DataStore. Can be set to TRUE if the user wants to work with the development environment on DataStore. Setting dev = TRUE will only work if the user has the appropriate permissions. Requires secure = TRUE.}
}
\value{
Logical.FALSE if the DataStore reference is not a data package. TRUE if the DataStore reference is a data package.
}
\description{
The function tests whether a valid DataStore reference is a data package or not. If it is a data package, the value "TRUE" is returned. If it is not a data package, the value FALSE is returned. The default setting assumes the user is logged on to the VPN (secure = TRUE), but this can be set to secure = FALSE for non-NPS users. This is a relatively simple helper function that will not test whether a reference exists and may produce unexpected errors if the reference does not exist or is not accessible. You are advised to run \code{check_reference_exists()} prior to using this function.
}
\examples{
\dontrun{
check_is_data_package(1234567)
check_is_data_package(1234567, secure = FALSE)
}
}
