% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_core_metadata.R
\name{.get_contacts}
\alias{.get_contacts}
\title{Extracts contacts and email addresses from EML metadata}
\usage{
.get_contacts(metadata)
}
\arguments{
\item{metadata}{an EML formatted R object}
}
\value{
dataframe
}
\description{
`.get_contacts()` extracts the "contacts" element from EML metadata and returns it as a dataframe with three columsn, first a column indicating that each row is an contact. Second, and column with the contact's name (first last). Third, the contact's email address.
}
\details{
`r lifecycle::badge('experimental')`
}
\examples{
\dontrun{
contacts <- get_contacts(metadata)
}
}
\keyword{private}
