% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_user_input}
\alias{.get_user_input}
\title{Get Binary User Input}
\usage{
.get_user_input()
}
\value{
Factor. 1 or 2.
}
\description{
Prompts for, gets, and returns binary user input (1 or 2)
}
\examples{
\dontrun{
var1 <- .get_user_input()
}
}
