% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceInfo.R
\name{DSReferenceInfo}
\alias{DSReferenceInfo}
\title{Get citation for Data Store holding info by HoldingID}
\usage{
DSReferenceInfo(HoldingID, field)
}
\arguments{
\item{HoldingID}{The six-seven digit reference / holding ID number unique to the data store record.}

\item{field}{is one of the following: "Title" returns the title of the data store reference as a string value;
"Abstract" returns the abstract as a string value; "Citation" returns the citation as a string value, and "Keywords" returns a vector containing
all keywords as character values.}
}
\description{
\code{DSReferenceInfo(HoldingID,field)} returns a character string or a vector with information from one of the
metadata fields in a Data Store reference's associated xml file.
}
\examples{
DSRefernceInfo(2266196,"Title")
}
