% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataPackage.R
\name{get.DataPackage}
\alias{get.DataPackage}
\title{Retrieve public digital data package holding from Data Store.}
\usage{
get.DataPackage(ReferenceID, Secure = FALSE)
}
\arguments{
\item{ReferenceID}{is a 6-7 digit number corresponding to the reference ID of the data package.}

\item{Secure}{logical TRUE (default) or FALSE indicating whether the file should be acquired using data services available to NPS internal staff only.}
}
\description{
getDataPackage downloads a data package from data store, copies it to the /data directory, and unzips it into a sub-directory of /data with the "HoldingID" name.
}
\examples{
\dontrun{
get.dataPackage(2272461,Secure=TRUE)
}
}
