% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParkUnitInfo.R
\name{get.parkCode}
\alias{get.parkCode}
\title{Restricts get.unitCode to just National Parks}
\usage{
get.parkCode(Park)
}
\arguments{
\item{Park}{is a case-insensitive string containing some part of the unit's FullName, e.g "Yellow".}
}
\value{
one data frame to the global environment. May contain multiple matches. Sufficient detail should be provided to choose the appropriate UnitCode for use with other NPSutils functions such as get.parkTaxonReferences (in ReferenceInfo.R). Dataframe contains UnitCode, FullName, UnitLifeCycle, Network, Region, and StateCodes.
}
\description{
\code{get.parkCode} is identical to \code{get.unitCode} except output is restricted to just National Parks (as opposed to including networks, etc.).Accesses info from irmaservices.nps.gov. Search for park or park unit with any string and return all applicable national park unitCodes. Handy for use with get.dataPackage if you don't know a Park's UnitCode.
}
\examples{
\dontrun{
get.parkCode("Yellow")
}
}
