% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReferenceInfo.R
\name{get.parkTaxonCitations}
\alias{get.parkTaxonCitations}
\title{Get DataStore Citations for a park-species combination}
\usage{
get.parkTaxonCitations(ParkCode, TaxonCode)
}
\arguments{
\item{ParkCode}{The four-letter unit code for the park of interest.}

\item{TaxonCode}{Taxonomic Serial Number for the taxon of interest.}
}
\value{
a dataframe
}
\description{
\code{get.parkTaxonCitations} returns a vector of citations in Data Store corresponding to a particular taxon at a park.

Note that this function only returns citations that have been "tagged" with a taxon code and may represent only a subset of references that have information about a taxon.
}
\examples{
\dontrun{
get.parkTaxonCitations("APIS", 126749)
}
}
