% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReferenceInfo.R
\name{get.parkTaxonRefs}
\alias{get.parkTaxonRefs}
\title{Get DS References for a park-species combination}
\usage{
get.parkTaxonRefs(ParkCode, TaxonCode)
}
\arguments{
\item{ParkCode}{The four-letter unit code for the park of interest.}

\item{TaxonCode}{Taxonomic Serial Number for the taxon of interest.}
}
\description{
\code{get.parkTaxonRefs} returns a data frame of metadata for references in Data Store corresponding to a particular taxon at a park.
The resultant data frame is then usable by other functions to extract metadata from the records.

Note that this function only returns references that have been "tagged" with a taxon code and may represent only a
subset of references that have information about a taxon.
}
\examples{
\dontrun{
get.parkTaxonRefs("APIS", 126749)
}
}
