% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParkUnitInfo.R
\name{get.unitInfo}
\alias{get.unitInfo}
\title{Search irmaservices.nps using any piece of information.}
\usage{
get.unitInfo(
  Code = NULL,
  Park = NULL,
  LifeCycle = NULL,
  NetworkCode = NULL,
  NetName = NULL,
  RegionAbb = NULL,
  Region = NULL,
  State = NULL
)
}
\arguments{
\item{Code}{defaults to NULL. Is a case-insensitive string. It typically is 4 letters long and typically does not include numbers but may be longer, shorter, or include special characters such as "-".}

\item{Park}{defaults to NULL. Is a case-insensitive string. It will search for any subset of the FullName or parks or park units}

\item{NetworkCode}{defaults to NULL. Is a case-insensitive string for the network code (or some subset of it). Note: not all park units have Network designations.}

\item{NetName}{defaults to NULL. Is a case-insensitive string containing of the network name (or some subset of it). Note not all park units are associated with networks.}

\item{RegionAbb}{defaults to NULL. Is a case-insensitive string containing the region abbreviation (or some subset of it), e.g.IMR for Intermountain Region. Not all park units are associated with a region.}

\item{Region}{defaults to NULL. Is a case-insensitive string of the region name (or some subset of it). Not all park units are associated with a region.}

\item{State}{defaults to NULL. Is a case-insensitive search of the park unit's state. States are indicated using 2-letter abbreviations. Park units may span multiple states. Not all parks have state designations.}

\item{LifCycle}{is a case-insensitive string that will filter search results based on LifeCycle status. Currently acceptable values are:
* "Active", "Inactive" or "Pending"}
}
\value{
one data frame to the global environment. May contain multiple matches. Sufficient detail should be provided to determine the intended  Park Unit's name.
}
\description{
\code{get.unitInfo} accesses info from irmaservices.nps.gov and allows you to search a Park Unit based on any number or combination of parameters. Not all parameters need to be specified, but it is probably worth specifying which parameters ARE specified, e.g. get.unitInfo(LifeCycle="Inactive"). If the arguments are not specified, they will default to the order supplied in the function.
}
\examples{
\dontrun{
get.UnitInfo(State="CO")
}
}
