% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataPackage.R
\name{get_data_package}
\alias{get_data_package}
\title{Retrieve public digital data package holding from Data Store.}
\usage{
get_data_package(reference_id, secure = FALSE)
}
\arguments{
\item{reference_id}{is a 6-7 digit number corresponding to the reference ID of the data package.}

\item{secure}{logical TRUE (default) or FALSE indicating whether the file should be acquired using data services available to NPS internal staff only.}
}
\description{
get_data_package downloads a data package from data store, copies it to the /data directory, and unzips it into a sub-directory of /data with the "HoldingID" name.
}
\examples{
\dontrun{
get_data_package(2272461, Secure = TRUE)
}
}
