% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_interactions.R
\name{get_new_version_id}
\alias{get_new_version_id}
\title{Get newest DataStore reference version}
\usage{
get_new_version_id(reference_id, secure = TRUE, dev = FALSE)
}
\arguments{
\item{reference_id}{Integer. A 7-digit number indicating the datastore reference to be queried.}

\item{secure}{Logical. Defaults to TRUE, indicating that the user is logged on to the NPS VPN (or is accessing the resource from an NPS office). Can be set to FALSE for non-NPS users. If \code{secure} is set to FALSE, \code{dev} is automatically set to FALSE.}

\item{dev}{Logical. Defaults to FALSE, indicating that the user wishes to work with the production version of DataStore. Can be set to TRUE if the user wants to work with the development environment on DataStore. Setting dev = TRUE will only work if the user has the appropriate permissions. Requires secure = TRUE.}
}
\value{
Integer. A 7-digit integer corresponding to the DataStore reference ID for the newest version of the supplied DataStore reference. If the supplied reference has not been versioned, NULL is returned.
}
\description{
If a DataStore reference has been versioned, \code{get_new_version_id()} will return the DataStore reference ID for the newest version. If the reference has not been versioned, the function returns NULL.
}
\examples{
\dontrun{
get_new_version_id(1234567)
get_new_version_id(1234567, secure = FALSE)
}
}
