% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReferenceInfo.R
\name{get_park_taxon_citations}
\alias{get_park_taxon_citations}
\title{Get DataStore Citations for a park-species combination}
\usage{
get_park_taxon_citations(park_code, taxon_code)
}
\arguments{
\item{park_code}{The four-letter unit code for the park of interest.}

\item{taxon_code}{Taxonomic Serial Number for the taxon of interest.}
}
\value{
a dataframe
}
\description{
\code{get_park_taxon_citations} returns a vector of citations in Data Store corresponding to a particular taxon at a park.

Note that this function only returns citations that have been "tagged" with a taxon code and may represent only a subset of references that have information about a taxon.
}
\examples{
\dontrun{
get_park_taxon_citations("APIS", 126749)
}
}
