% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReferenceInfo.R
\name{get_park_taxon_url}
\alias{get_park_taxon_url}
\title{Get URL for references for a park-species combination}
\usage{
get_park_taxon_url(park_code, taxon_code)
}
\arguments{
\item{park_code}{The four-letter unit code for the park of interest.}

\item{taxon_code}{The IRMA taxon code.}
}
\description{
\code{get_park_taxon_url} returns a vector of URLs for references in Data Store corresponding to a particular taxon at a park.
}
\details{
Note that this function only returns URLs for references that have been "tagged" with a taxon code and may represent only a
subset of references that have information about a taxon.
}
\examples{
\dontrun{
get_park_taxon_url("APIS", 126749)
}
}
