% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analyses.R
\name{get_refs_info}
\alias{get_refs_info}
\title{Return Basic information about a list of DataStore References}
\usage{
get_refs_info(
  reference_type = "dataPackage",
  no_of_entries = 500,
  secure = FALSE
)
}
\arguments{
\item{reference_type}{String. The reference type to to query data store for. Defaults to data package ("dataPackage").}

\item{no_of_entries}{Integer. The number of entries to return per page (where only one "page" of results is returned by default). Defaults to 500.}

\item{secure}{Logical. Defaults to FALSE for external users. Setting secure = TRUE will, with the proper credentials, return DataStore references with visibility set to both Public and Restricted.}
}
\value{
a data frame
}
\description{
The function will return a data frame containing information about a given number of references within a reference type. The data returned includes the reference ID (referenceId), the date the references was activated on DataStore (dateOfIssue), the references visibility (visibility), the number of files associated with the reference (fileCount), the access level of the files (fileAccess), the reference title (title), the abbreviated citation (citation), the URL for the DataStore reference (referenceUrl), the group-type for the reference (referenceGroupType), the type of reference (typeName), whether the reference has a DOI associated with it (isDOI), whether their is a newer version of the reference (newVersion) and what the most recent version of the reference is (mostRecentReference).
}
\examples{
\dontrun{
get_ref_info()
}

}
