% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataPackage.R
\name{getDataPackages}
\alias{getDataPackages}
\title{Retrieve digital data package holding from DataStore.}
\usage{
getDataPackages(reference_id, secure = FALSE, path = here::here())
}
\arguments{
\item{reference_id}{is a 6-7 digit number corresponding to the reference ID of the data package.}

\item{secure}{logical indicating whether the file should be acquired using data services available to NPS internal staff only. Defaults to FALSE for public data. TRUE indicates internal data and requires a VPN connection (unless you are in an NPS office).}

\item{path}{String. Indicates the location that the "data" directory and all subdirectories and files should be written to. Defaults to the working directory.}
}
\value{
String. The path where the /data folder is written and all data package sub-directories and data files are contained.
}
\description{
`getDataPackages()` creates a directory called "data" in the current working directory (unless it already exists). For each data package, it writes a new sub-directory of "data" named with the corresponding data package reference ID. All the data package files are then copied to that directory.
}
\examples{
\dontrun{
getDataPackages(2272461, secure = FALSE)
pathToDataForPiping <- getDataPackages(2272461, secure = TRUE)
}
}
