% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParkUnitInfo.R
\name{getParkCode}
\alias{getParkCode}
\title{\code{getParkCode} is identical to \code{getUnitCode} except output is restricted to just National Parks (as opposed to including networks, etc.).Accesses info from irmaservices.nps.gov. Search for park or park unit with any string and return all applicable national park unitCodes. Handy for use with GetDataPackage if you don't know a Park's UnitCode.}
\usage{
getParkCode(Park)
}
\arguments{
\item{Park}{is a case-insensitive string containing some part of the unit's FullName.}
}
\value{
one data frame to the global environment. May contain multiple matches. Sufficient detail should be provided to choose the appropriate UnitCode for use with other DSTools functions such as DSGetParkTaxonRefernces (in ReferenceInfo.R). Dataframe contains UnitCode, FullName, UnitLifeCycle, Network, Region, and StateCodes.
}
\description{
\code{getParkCode} is identical to \code{getUnitCode} except output is restricted to just National Parks (as opposed to including networks, etc.).Accesses info from irmaservices.nps.gov. Search for park or park unit with any string and return all applicable national park unitCodes. Handy for use with GetDataPackage if you don't know a Park's UnitCode.
}
