% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReferenceInfo.R
\name{getParkTaxonURLs}
\alias{getParkTaxonURLs}
\title{Get URL for references for a park-species combination}
\usage{
getParkTaxonURLs(ParkCode, TaxonCode)
}
\arguments{
\item{ParkCode}{The four-letter unit code for the park of interest.}

\item{TaxonCode}{Taxonomic Serial Number for the taxon of interest.}
}
\description{
\code{DSgetParkTaxonReferenceURL} returns a vector of URLs for references in Data Store corresponding to a particular taxon at a park.
}
\details{
Note that this function only returns URLs for references that have been "tagged" with a taxon code and may represent only a
subset of references that have information about a taxon.
}
\examples{
getParkTaxonURLs("APIS", 126749)
}
