% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDomains.R
\name{load.domains}
\alias{load.domains}
\title{Read contents of data package file and construct a data frame based on the metadata file summarizing the enumerated domains for
categorical fields and their types/definitions.}
\usage{
load.domains(HoldingID, dataFormat, metadataFormat)
}
\arguments{
\item{HoldingID}{is a 6-7 digit number corresponding to the holding ID of the data package zip file.}

\item{dataFormat}{is a character value indicating the format of the data set(s) within the data package. Currently
allowable options are:
* "csv" for comma separated value text files
* "gdb" for file geodatabases}

\item{metadataFormat}{is a character value indicating the format of the metadata file within the data package.
Currently allowable options are:
* "eml" for eml-compliant xml metadata files
* "fgdc" for FGDC-compliant xml metadata files}
}
\value{
one data frame to the global environment.
}
\description{
\code{load.domains} reads the data file from a package and loads a list of fields and their attributes into a data frame.
}
\examples{
\dontrun{
load.domains(2266200,dataFormat="gdb",metadataFormat="fgdc")
}
}
