% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMetadata.R
\name{load.metadata}
\alias{load.metadata}
\title{Read contents of data package file and construct a data frame based on the metadata file summarizing the fields and their types/definitions.}
\usage{
load.metadata(HoldingID)
}
\arguments{
\item{HoldingID}{is a 6-7 digit number corresponding to the holding ID of the data package zip file.}
}
\value{
one data frame to the global environment.
}
\description{
\code{load.metadata} reads the metadata file from a previously downloaded package and loads a list of fields and their attributes into a data frame.
}
\examples{
\dontrun{
load.metadata(2266200)
}
}
