% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_package.R
\name{load_data_package_deprecated}
\alias{load_data_package_deprecated}
\title{Read contents of data package and constructs a list of tibbles based on the data file(s)}
\usage{
load_data_package_deprecated(reference_id)
}
\arguments{
\item{reference_id}{is a 6-7 digit number corresponding to the reference ID of the data package.}
}
\value{
a list of one or more tibbles contained within the data package to the global environment.
}
\description{
`load_data_package_deprecated()` reads the data file(s) from a package and loads it into a list of tibbles. Current implementation only supports .csv data files.
}
\details{
`r lifecycle::badge("deprecated")`
}
\examples{
\dontrun{
load_data_package_deprecated(2272461)
}
}
