% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataPackageList.R
\name{load_data_package_list}
\alias{load_data_package_list}
\title{Read contents of data package file and construct a vector with the list of data frames to be created.}
\usage{
load_data_package_list(holding_id, data_format, metadata_format)
}
\arguments{
\item{holding_id}{is a 6-7 digit number corresponding to the holding ID of the data package zip file.}

\item{data_format}{is a character value indicating the format of the data set(s) within the data package. Currently
allowable options are:
* "csv" for comma separated value text files
* "gdb" for file geodatabases}

\item{metadata_format}{is a character value indicating the format of the metadata file within the data package.
Currently allowable options are:
* "eml" for eml-compliant xml metadata files
* "fgdc" for FGDC-compliant xml metadata files}
}
\value{
one or more data frames contained within the data package to the global environment.
}
\description{
\code{load_data_package_list} reads the data file from a package and loads it into a data frame.
}
\examples{
\dontrun{
load_data_package_list(2272461, data_format = "csv", metadata_format = "eml")
}

}
