% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_domains.R
\name{load_domains}
\alias{load_domains}
\title{Read contents of data package file and construct a data frame based on the metadata file summarizing the enumerated domains for
categorical fields and their types/definitions.}
\usage{
load_domains(holding_id, data_format, metadata_format)
}
\arguments{
\item{holding_id}{is a 6-7 digit number corresponding to the holding ID of the data package zip file.}

\item{data_format}{is a character value indicating the format of the data set(s) within the data package. Currently
allowable options are:
\itemize{
\item "csv" for comma separated value text files
\item "gdb" for file geodatabases
}}

\item{metadata_format}{is a character value indicating the format of the metadata file within the data package.
Currently allowable options are:
\itemize{
\item "eml" for eml-compliant xml metadata files
\item "fgdc" for FGDC-compliant xml metadata files
}}
}
\value{
one data frame to the global environment.
}
\description{
\code{load_domains} reads the data file from a package and loads a list of fields and their attributes into a data frame.
}
\examples{
\dontrun{
load_domains(2266200, data_format = "gdb", metadata_format = "fgdc")
}
}
