% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataPackageList.R
\name{loadDataPackageList}
\alias{loadDataPackageList}
\title{Read contents of data package file and construct a vector with the list of data frames to be created.
For file geodatabases this will only extract spatial data features.}
\usage{
loadDataPackageList(HoldingID, dataFormat, metadataFormat)
}
\arguments{
\item{HoldingID}{is a 6-7 digit number corresponding to the holding ID of the data package zip file.}

\item{dataFormat}{is a character value indicating the format of the data set(s) within the data package. Currently
allowable options are:
* "csv" for comma separated value text files
* "gdb" for file geodatabases}

\item{metadataFormat}{is a character value indicating the format of the metadata file within the data package.
Currently allowable options are:
* "eml" for eml-compliant xml metadata files
* "fgdc" for FGDC-compliant xml metadata files}
}
\value{
one or more data frames contained within the data package to the global environment.
}
\description{
\code{load.dataPackageList} reads the data file from a package and loads it into a data frame.
}
\examples{
\dontrun{
load.dataPackageList(2272461, dataFormat="csv", metadataFormat="eml")
}
}
