% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMetadata.R
\name{loadMetadata}
\alias{loadMetadata}
\title{Read contents of data package file and construct a data frame based on the metadata file summarizing the fields and their types/definitions.}
\usage{
loadMetadata(HoldingID)
}
\arguments{
\item{HoldingID}{is a 6-7 digit number corresponding to the holding ID of the data package zip file.}
}
\value{
one data frame to the global environment.
}
\description{
\code{loadMetadata} reads the metadata file from a previously downloaded package and loads a list of fields and their attributes into a data frame.
}
\examples{

attributeTable<-loadMetadata(2266200)
}
