% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPSutils-package.R
\docType{package}
\name{NPSutils-package}
\alias{NPSutils}
\alias{NPSutils-package}
\title{NPSutils: Collection of Functions to read and manipulate information from the NPS DataStore}
\description{
NPSutils is a collection of functions for interacting with NPS DataStore repository.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nationalparkservice/NPSutils}
  \item Report bugs at \url{https://github.com/nationalparkservice/NPSutils/issues}
}

}
\author{
\strong{Maintainer}: Robert Baker \email{robert_baker@nps.gov} (\href{https://orcid.org/0000-0001-7591-5035}{ORCID})

Authors:
\itemize{
  \item Joe DeVivo
  \item Judd Patterson \email{Judd_Patterson@nps.gov} (\href{https://orcid.org/0000-0002-0951-7917}{ORCID})
}

}
\keyword{internal}
