% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateDataPackage.R
\name{validate.dataPackage}
\alias{validate.dataPackage}
\title{Validate a Tabular Data Package}
\usage{
validate.dataPackage(File)
}
\arguments{
\item{File}{is a path to a zipped data package}
}
\description{
\code{validateDataPackage} checks a zipped data package for required files and structure.
}
\examples{
\dontrun{
validate.dataPackage(File="C:/test/test_package.zip")
}
}
