% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateDataPackage.R
\name{validate_data_package}
\alias{validate_data_package}
\title{Validate a Tabular Data Package}
\usage{
validate_data_package(file)
}
\arguments{
\item{file}{is a path to a zipped data package}
}
\description{
\code{validate_data_package} checks a zipped data package for required files and structure. This function is still under construction.
}
\examples{
\dontrun{
validate_data_package(File = "C:/test/test_package.zip")
}
}
