% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReferenceInfo.R, R/meta_analyses.R
\name{get_ref_info}
\alias{get_ref_info}
\title{Get citation for Data Store holding info by HoldingID}
\usage{
get_ref_info(
  reference_type = "dataPackage",
  no_of_entries = 500,
  secure = FALSE
)

get_ref_info(
  reference_type = "dataPackage",
  no_of_entries = 500,
  secure = FALSE
)
}
\arguments{
\item{reference_type}{String. Defaults to "dataPackage". The reference type to query DataStore for.}

\item{no_of_entries}{Integer. Defaults to 500. The number of entries to return.}

\item{secure}{Logical. Defaults to FALSE. Should the function use a secure API or the pubic API?}

\item{holding_id}{The six-seven digit reference / holding ID number unique to the data store record.}

\item{field}{is one of the following: "Title" returns the title of the data store reference as a string value;
"Abstract" returns the abstract as a string value; "Citation" returns the citation as a string value, and "Keywords" returns a vector containing
all keywords as character values.}
}
\value{
a data frame
}
\description{
\code{get_ref_info} returns a character string or a vector with information from one of the
metadata fields in a Data Store reference's associated xml file.

The function will return a data frame containing information about a given number of references within a reference type. The data returned includes the reference ID (referenceId), the date the references was activated on DataStore (dateOfIssue), the references visibility (visibility), the number of files associated with the reference (fileCount), the access level of the files (fileAccess), the reference title (title), the abbreviated citation (citation), the URL for the DataStore reference (referenceUrl), the group-type for the reference (referenceGroupType), the type of reference (typeName), whether the reference has a DOI associated with it (isDOI), whether their is a newer version of the reference (newVersion) and what the most recent version of the reference is (mostRecentReference).
}
\examples{
\dontrun{
get_ref_info(2266196, "Title")
}
\dontrun{
get_ref_info()
}

}
