% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParkUnitInfo.R
\name{getUnitCode}
\alias{getUnitCode}
\title{#' To do:
Warnings and checks on parameter inputs are not yet implemented.
The output dataframe is a little unwieldy and could be cleaned up.
Dynamically access NPS park unit code data}
\usage{
getUnitCode(Unit)
}
\arguments{
\item{Unit}{is a case-insensitive string containing some part of the unit's FullName.}
}
\value{
one data frame to the global environment. May contain multiple matches. Sufficient detail should be provided to choose the appropriate UnitCode for use with other DSTools functions such as getParkTaxonReferences (in getReferenceInfo.R).
}
\description{
allows the user to access information based on park codes 

\code{getUnitCode} accesses info from irmaservices.nps.gov. Search for park or park unit with any string and return all applicable UnitCodes. Handy for use with GetDataPackage if you don't know a Park's UnitCode.
}
