% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataPackage.R
\name{load_data_package}
\alias{load_data_package}
\title{Read contents of data package file and constructs a list of tibbles based on the data file(s)}
\usage{
load_data_package(holding_id)
}
\arguments{
\item{holding_id}{is a 6-7 digit number corresponding to the holding ID of the data package.}
}
\value{
a list of one or more tibbles contained within the data package to the global environment.
}
\description{
\code{load_data_package} reads the data file(s) from a package and loads it into a list of tibbles. Current implementation only supports .csv data files. Planned future iterations will support .json and .xml data files.
}
\examples{
\dontrun{
load_data_package(2272461)
}
}
