% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataPackage.R
\name{loadDataPackage}
\alias{loadDataPackage}
\title{Read contents of data packge file and construct a data frame based on the data file(s) and associated metadata.}
\usage{
loadDataPackage(HoldingID, dataFormat, metadataFormat, features = NULL)
}
\arguments{
\item{HoldingID}{is a 6-7 digit number corresponding to the holding ID of the data package zip file.}

\item{dataFormat}{is a character value indicating the format of the data set(s) within the data package. Currently
allowable options are:
* "csv" for comma seperated value text files
* "gdb" for file geodatabases}

\item{metadataFormat}{is a character value indicating the format of the metadata file within the data package.
Currently allowable options are:
* "eml" for eml-compliant xml metadata files
* "fgdc" for FGDC-compliant xml metadata files}

\item{features}{is a character value indicating the name of the feature class / data set (only used with file geodatabases)}
}
\value{
one or more data frames contained within the data package to the global enviroment.
}
\description{
\code{loadDataPackage} reads the data file from a package and loads it into one or more data frames.
}
\examples{

datasetList<-loadDataPackage(2266200,dataFormat="gdb",metadataFormat="fgdc",features=NULL)
}
